requirements.txt
alembic==1.11.1
blinker==1.6.2
build==0.10.0
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pyproject_hooks==1.0.0
python-slugify==8.0.1
six==1.16.0
SQLAlchemy==2.0.16
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.6.3
Werkzeug==2.3.6
WTForms==3.0.1
